"Start and finish"
I've placed the camera at the right starting-place for the test-level.
I've also made an "F"-model which spins at the goal-point, or
"finish"-point.
You can also toggle on wireframe-view with the "F"-key on your keyboard,
but you have to restart the game for it to toggle off again.