Fullscreen!
Added a function for making the game go into fullscreen-mode. This is
done by pressing the "F11"-key. You can use the same key to turn it back
into windowed-mode.

Also made it possible to toggle on/off wireframe-mode as you wish.
Additionally, "keys" in the Keyboard-class now has three different
possible states in stead of just two. "KEY_PRESSED", "KEY_RELEASED" and
"KEY_UNCHANGED". When a key-state is checked it's state will be reset to
"KEY_UNCHANGED".