Version 0.1.1a
I've done some more changes to the input-classes. Now a key/button has
one of four possible states. PRESSED, UNCHANGED_FROM_PRESS, RELEASED and
UNCHANGED_FROM_RELEASE. This makes it possible to either listen to a
singular "press" or to check if a key/button is held down, all with the
same system!

I've also made it possible to toggle VSync on/off with the "V"-key.